+++ animation/src/animation/Scale.java
@@ -4,8 +4,8 @@ package animation;
  * Represents a Scale transformation, extends methods from the AbstractTransformation.
  */
 public class Scale extends AbstractTransformation {
-  final float toX;
-  final float toY;
+  final int toX;
+  final int toY;
   
   /**
    * Constructor calls the AbstractTransformation super class, takes in the color to change to.
@@ -14,7 +14,7 @@ public class Scale extends AbstractTransformation {
    * @param timeStart (int) start time for the transformation
    * @param timeEnd (int) start time for the transformation
    */
-  public Scale(String tid, float toX, float toY, int timeStart, int timeEnd) {
+  public Scale(String tid, int toX, int toY, int timeStart, int timeEnd) {
     super(tid, timeStart, timeEnd);
     if (toX <= 0 || toY <= 0) {
       throw new IllegalArgumentException("Length must be positive integer or 0.");
@@ -27,12 +27,12 @@ public class Scale extends AbstractTransformation {
   
   //GETTERS----------------------------------------------------------------------------------------
   @Override
-  public float getToX() {
+  public int getToX() {
     return this.toX;
   }
   
   @Override
-  public float getToY() {
+  public int getToY() {
     return this.toY;
   }
